+ initial Vector

+++ src/Apogey.h
@@ -63,8 +63,8 @@ class ApogeyRenderer : public Crt8275Renderer
     protected:
         const uint8_t* getCurFontPtr(bool gpa0, bool gpa1, bool hglt) override;
         const uint8_t* getAltFontPtr(bool gpa0, bool gpa1, bool hglt) override;
-        uint32_t getCurFgColor(bool gpa0, bool gpa1, bool hglt) override;
-        uint32_t getCurBgColor(bool gpa0, bool gpa1, bool hglt) override;
+        uint8_t getCurFgColor(bool gpa0, bool gpa1, bool hglt) override;
+        uint8_t getCurBgColor(bool gpa0, bool gpa1, bool hglt) override;
         wchar_t getUnicodeSymbol(uint8_t chr, bool gpa0, bool gpa1, bool hglt) override;
 
     private:
