+++ src/Partner.h
@@ -207,8 +207,8 @@ class PartnerRenderer : public Crt8275Renderer
     protected:
         const uint8_t* getCurFontPtr(bool gpa0, bool gpa1, bool hglt) override;
         const uint8_t* getAltFontPtr(bool gpa0, bool gpa1, bool hglt) override;
-        uint32_t getCurFgColor(bool gpa0, bool gpa1, bool hglt) override;
-        uint32_t getCurBgColor(bool gpa0, bool gpa1, bool hglt) override;
+        uint8_t getCurFgColor(bool gpa0, bool gpa1, bool hglt) override;
+        uint8_t getCurBgColor(bool gpa0, bool gpa1, bool hglt) override;
         wchar_t getUnicodeSymbol(uint8_t chr, bool gpa0, bool gpa1, bool hglt) override;
 
     private:
@@ -239,9 +239,9 @@ class PartnerMcpgRenderer : public Crt8275Renderer
         static EmuObject* create(const EmuValuesList&) {return new PartnerMcpgRenderer();}
 
     protected:
-        uint32_t getCurFgColor(bool gpa0, bool gpa1, bool hglt) override;
-        uint32_t getCurBgColor(bool gpa0, bool gpa1, bool hglt) override;
-        void customDrawSymbolLine(uint32_t* linePtr, uint8_t symbol, int line, bool lten, bool vsp, bool rvv, bool gpa0, bool gpa1, bool hglt) override;
+        uint8_t getCurFgColor(bool gpa0, bool gpa1, bool hglt) override;
+        uint8_t getCurBgColor(bool gpa0, bool gpa1, bool hglt) override;
+        void customDrawSymbolLine(uint8_t* linePtr, uint8_t symbol, int line, bool lten, bool vsp, bool rvv, bool gpa0, bool gpa1, bool hglt) override;
 
     private:
         const uint8_t* m_fontPtr = nullptr;
